Second Stage
A new custom user model is created, as well as a second view, the
rent view.

User authentication is done using tokens. In order to post, patch
and delete from the Movies view, the user must be a staff member.
Tokens are passed through the `curl` command, and are created
for an already existing user by running
`./manage.py drf_create_token <username>` .

Currently, if an authenticated user sends a GET request to
/api/rent/<movie id>, this movie id is added to his movies_rented
list.