Third Stage
Create an explicit RentMovie model, which is the through table
for the many-to-many relation between User and Movie. RentMovie
keeps track of the status, rent_date, updated_date and cost.

In order to rent and return a movie, the user must now send a POST
request to /rent and /return, specifying the movie id in the
request's data.

Also, in order to filter the movies in the movies view based on
the category, instead of going to /movies/<category>, it is now
a GET argument. So for example
`GET http://127.0.0.1:8000/api/movies?category=Horror`