Clear up the difference between get_cost(True) and get_cost(False)
If the movie has already been returned, get_cost(True) is used
in order to calculate the cost between the date it was rented
and the date it was returned. This happens in UserProfileView,
when listing already returned rented movies.

If the movie has not been returned yet, or is going to be returned
now by a POST request to /return, get_cost(False) is used. Now the
cost will be calculated using the current date.